@@ -1,9 +1,10 @@
-//go:build (linux || (windows && !arm) || darwin) && (386 || amd64 || arm || arm64) && !race
-// +build linux windows,!arm darwin
+//go:build (linux || (windows && !arm && !arm64) || darwin) && (386 || amd64 || arm || arm64) && !race && !nosilk
+// +build linux windows,!arm,!arm64 darwin
 // +build 386 amd64 arm arm64
 // +build !race
+// +build !nosilk
 
-package codec
+package silk
 
 import (
 	"os"
@@ -12,12 +13,14 @@ import (
 
 	"github.com/pkg/errors"
 	"github.com/wdvxdr1123/go-silk"
+
+	"github.com/Mrs4s/go-cqhttp/internal/base"
 )
 
 const silkCachePath = "data/cache"
 
-// EncodeToSilk 将音频编码为Silk
-func EncodeToSilk(record []byte, tempName string, useCache bool) (silkWav []byte, err error) {
+// encode 将音频编码为Silk
+func encode(record []byte, tempName string) (silkWav []byte, err error) {
 	// 1. 写入缓存文件
 	rawPath := path.Join(silkCachePath, tempName+".wav")
 	err = os.WriteFile(rawPath, record, os.ModePerm)
@@ -29,7 +32,7 @@ func EncodeToSilk(record []byte, tempName string, useCache bool) (silkWav []byte
 	// 2.转换pcm
 	pcmPath := path.Join(silkCachePath, tempName+".pcm")
 	cmd := exec.Command("ffmpeg", "-i", rawPath, "-f", "s16le", "-ar", "24000", "-ac", "1", pcmPath)
-	if Debug {
+	if base.Debug {
 		cmd.Stdout = os.Stdout
 		cmd.Stderr = os.Stderr
 	}
@@ -47,15 +50,13 @@ func EncodeToSilk(record []byte, tempName string, useCache bool) (silkWav []byte
 	if err != nil {
 		return nil, errors.Wrap(err, "silk encode error")
 	}
-	if useCache {
-		silkPath := path.Join(silkCachePath, tempName+".silk")
-		err = os.WriteFile(silkPath, silkWav, 0o666)
-	}
+	silkPath := path.Join(silkCachePath, tempName+".silk")
+	err = os.WriteFile(silkPath, silkWav, 0o666)
 	return
 }
 
-// RecodeTo24K 将silk重新编码为 24000 bit rate
-func RecodeTo24K(data []byte) []byte {
+// resample 将silk重新编码为 24000 bit rate
+func resample(data []byte) []byte {
 	pcm, err := silk.DecodeSilkBuffToPcm(data, 24000)
 	if err != nil {
 		panic(err)
