@@ -3,13 +3,13 @@ package coolq
 import (
 	"crypto/md5"
 	"encoding/hex"
+	"encoding/json"
 	"fmt"
 	"math"
 	"os"
 	"path"
 	"path/filepath"
 	"runtime"
-	"runtime/debug"
 	"strconv"
 	"strings"
 	"time"
@@ -17,51 +17,222 @@ import (
 	"github.com/Mrs4s/MiraiGo/binary"
 	"github.com/Mrs4s/MiraiGo/client"
 	"github.com/Mrs4s/MiraiGo/message"
+	"github.com/Mrs4s/MiraiGo/utils"
 	log "github.com/sirupsen/logrus"
 	"github.com/tidwall/gjson"
 
+	"github.com/Mrs4s/go-cqhttp/db"
 	"github.com/Mrs4s/go-cqhttp/global"
+	"github.com/Mrs4s/go-cqhttp/internal/base"
+	"github.com/Mrs4s/go-cqhttp/internal/cache"
+	"github.com/Mrs4s/go-cqhttp/internal/param"
+	"github.com/Mrs4s/go-cqhttp/modules/filter"
 )
-
-// Version go-cqhttp的版本信息，在编译时使用ldflags进行覆盖
-var Version = "unknown"
-
-func init() {
-	if Version != "unknown" {
-		return
-	}
-	info, ok := debug.ReadBuildInfo()
-	if ok {
-		Version = info.Main.Version
-	}
-}
 
 // CQGetLoginInfo 获取登录号信息
 //
 // https://git.io/Jtz1I
-func (bot *CQBot) CQGetLoginInfo() MSG {
-	return OK(MSG{"user_id": bot.Client.Uin, "nickname": bot.Client.Nickname})
+// @route(get_login_info)
+func (bot *CQBot) CQGetLoginInfo() global.MSG {
+	return OK(global.MSG{"user_id": bot.Client.Uin, "nickname": bot.Client.Nickname})
 }
 
 // CQGetQiDianAccountInfo 获取企点账号信息
-func (bot *CQBot) CQGetQiDianAccountInfo() MSG {
+// @route(qidian_get_account_info)
+func (bot *CQBot) CQGetQiDianAccountInfo() global.MSG {
 	if bot.Client.QiDian == nil {
 		return Failed(100, "QIDIAN_PROTOCOL_REQUEST", "请使用企点协议")
 	}
-	return OK(MSG{
+	return OK(global.MSG{
 		"master_id":   bot.Client.QiDian.MasterUin,
 		"ext_name":    bot.Client.QiDian.ExtName,
 		"create_time": bot.Client.QiDian.CreateTime,
 	})
 }
 
+// CQGetGuildServiceProfile 获取频道系统个人资料
+// @route(get_guild_service_profile)
+func (bot *CQBot) CQGetGuildServiceProfile() global.MSG {
+	return OK(global.MSG{
+		"nickname":   bot.Client.GuildService.Nickname,
+		"tiny_id":    bot.Client.GuildService.TinyId,
+		"avatar_url": bot.Client.GuildService.AvatarUrl,
+	})
+}
+
+// CQGetGuildList 获取已加入的频道列表
+// @route(get_guild_list)
+func (bot *CQBot) CQGetGuildList() global.MSG {
+	fs := make([]global.MSG, 0, len(bot.Client.GuildService.Guilds))
+	for _, info := range bot.Client.GuildService.Guilds {
+		/* 做成单独的 api 可能会好些?
+		channels := make([]global.MSG, 0, len(info.Channels))
+		for _, channel := range info.Channels {
+			channels = append(channels, global.MSG{
+				"channel_id":   channel.ChannelId,
+				"channel_name": channel.ChannelName,
+				"channel_type": channel.ChannelType,
+			})
+		}
+		*/
+		fs = append(fs, global.MSG{
+			"guild_id":         info.GuildId,
+			"guild_name":       info.GuildName,
+			"guild_display_id": info.GuildCode,
+			// "channels":         channels,
+		})
+	}
+	return OK(fs)
+}
+
+// CQGetGuildMetaByGuest 通过访客权限获取频道元数据
+// @route(get_guild_meta_by_guest)
+func (bot *CQBot) CQGetGuildMetaByGuest(guildID uint64) global.MSG {
+	meta, err := bot.Client.GuildService.FetchGuestGuild(guildID)
+	if err != nil {
+		log.Errorf("获取频道元数据时出现错误: %v", err)
+		return Failed(100, "API_ERROR", err.Error())
+	}
+	return OK(global.MSG{
+		"guild_id":         meta.GuildId,
+		"guild_name":       meta.GuildName,
+		"guild_profile":    meta.GuildProfile,
+		"create_time":      meta.CreateTime,
+		"max_member_count": meta.MaxMemberCount,
+		"max_robot_count":  meta.MaxRobotCount,
+		"max_admin_count":  meta.MaxAdminCount,
+		"member_count":     meta.MemberCount,
+		"owner_id":         meta.OwnerId,
+	})
+}
+
+// CQGetGuildChannelList 获取频道列表
+// @route(get_guild_channel_list)
+func (bot *CQBot) CQGetGuildChannelList(guildID uint64, noCache bool) global.MSG {
+	guild := bot.Client.GuildService.FindGuild(guildID)
+	if guild == nil {
+		return Failed(100, "GUILD_NOT_FOUND")
+	}
+	if noCache {
+		channels, err := bot.Client.GuildService.FetchChannelList(guildID)
+		if err != nil {
+			log.Errorf("获取频道 %v 子频道列表时出现错误: %v", guildID, err)
+			return Failed(100, "API_ERROR", err.Error())
+		}
+		guild.Channels = channels
+	}
+	channels := make([]global.MSG, 0, len(guild.Channels))
+	for _, c := range guild.Channels {
+		channels = append(channels, convertChannelInfo(c))
+	}
+	return OK(channels)
+}
+
+// CQGetGuildMembers 获取频道成员列表
+// @route(get_guild_members)
+func (bot *CQBot) CQGetGuildMembers(guildID uint64) global.MSG {
+	guild := bot.Client.GuildService.FindGuild(guildID)
+	if guild == nil {
+		return Failed(100, "GUILD_NOT_FOUND")
+	}
+	members := make([]global.MSG, len(guild.Members))
+	bots := make([]global.MSG, len(guild.Bots))
+	admins := make([]global.MSG, len(guild.Admins))
+	for i, m := range guild.Members {
+		members[i] = convertGuildMemberInfo(m)
+	}
+	for i, m := range guild.Bots {
+		bots[i] = convertGuildMemberInfo(m)
+	}
+	for i, m := range guild.Admins {
+		admins[i] = convertGuildMemberInfo(m)
+	}
+	return OK(global.MSG{
+		"members": members,
+		"bots":    bots,
+		"admins":  admins,
+	})
+}
+
+// CQGetGuildRoles 获取频道角色列表
+// @route(get_guild_roles)
+func (bot *CQBot) CQGetGuildRoles(guildID uint64) global.MSG {
+	roles, err := bot.Client.GuildService.GetGuildRoles(guildID)
+	if err != nil {
+		log.Errorf("获取频道 %v 角色列表时出现错误: %v", guildID, err)
+		return Failed(100, "API_ERROR", err.Error())
+	}
+	return OK(global.MSG{
+		"roles": roles,
+	})
+}
+
+// CQCreateGuildRole 创建频道角色
+// @route(create_guild_role)
+func (bot *CQBot) CQCreateGuildRole(guildID uint64, name string, color uint32, independent bool, initialUsers gjson.Result) global.MSG {
+	userSlice := []uint64{}
+	if initialUsers.IsArray() {
+		for _, user := range initialUsers.Array() {
+			userSlice = append(userSlice, user.Uint())
+		}
+	}
+	role, err := bot.Client.GuildService.CreateGuildRole(guildID, name, color, independent, userSlice)
+	if err != nil {
+		log.Errorf("创建频道 %v 角色时出现错误: %v", guildID, err)
+		return Failed(100, "API_ERROR", err.Error())
+	}
+	return OK(global.MSG{
+		"role": role,
+	})
+}
+
+// CQDeleteGuildRole 删除频道角色
+// @route(delete_guild_role)
+func (bot *CQBot) CQDeleteGuildRole(guildID uint64, roleID uint64) global.MSG {
+	err := bot.Client.GuildService.DeleteGuildRole(guildID, roleID)
+	if err != nil {
+		log.Errorf("删除频道 %v 角色时出现错误: %v", guildID, err)
+		return Failed(100, "API_ERROR", err.Error())
+	}
+	return OK(nil)
+}
+
+// CQSetGuildMemberRole 设置用户在频道中的角色
+// @route(set_guild_member_role)
+func (bot *CQBot) CQSetGuildMemberRole(guildID uint64, set bool, roleID uint64, users gjson.Result) global.MSG {
+	userSlice := []uint64{}
+	if users.IsArray() {
+		for _, user := range users.Array() {
+			userSlice = append(userSlice, user.Uint())
+		}
+	}
+	err := bot.Client.GuildService.SetUserRoleInGuild(guildID, set, roleID, userSlice)
+	if err != nil {
+		log.Errorf("设置用户在频道 %v 中的角色时出现错误: %v", guildID, err)
+		return Failed(100, "API_ERROR", err.Error())
+	}
+	return OK(nil)
+}
+
+// CQModifyRoleInGuild 修改频道角色
+// @route(modify_role_in_guild)
+func (bot *CQBot) CQModifyRoleInGuild(guildID uint64, roleID uint64, name string, color uint32, indepedent bool) global.MSG {
+	err := bot.Client.GuildService.ModifyRoleInGuild(guildID, roleID, name, color, indepedent)
+	if err != nil {
+		log.Errorf("修改频道 %v 角色时出现错误: %v", guildID, err)
+		return Failed(100, "API_ERROR", err.Error())
+	}
+	return OK(nil)
+}
+
 // CQGetFriendList 获取好友列表
 //
 // https://git.io/Jtz1L
-func (bot *CQBot) CQGetFriendList() MSG {
-	fs := make([]MSG, 0, len(bot.Client.FriendList))
+// @route(get_friend_list)
+func (bot *CQBot) CQGetFriendList() global.MSG {
+	fs := make([]global.MSG, 0, len(bot.Client.FriendList))
 	for _, f := range bot.Client.FriendList {
-		fs = append(fs, MSG{
+		fs = append(fs, global.MSG{
 			"nickname": f.Nickname,
 			"remark":   f.Remark,
 			"user_id":  f.Uin,
@@ -70,10 +241,52 @@ func (bot *CQBot) CQGetFriendList() MSG {
 	return OK(fs)
 }
 
+// CQGetUnidirectionalFriendList 获取单向好友列表
+//
+// @route(get_unidirectional_friend_list)
+func (bot *CQBot) CQGetUnidirectionalFriendList() global.MSG {
+	list, err := bot.Client.GetUnidirectionalFriendList()
+	if err != nil {
+		log.Errorf("获取单向好友列表时出现错误: %v", err)
+		return Failed(100, "API_ERROR", err.Error())
+	}
+	fs := make([]global.MSG, 0, len(list))
+	for _, f := range list {
+		fs = append(fs, global.MSG{
+			"nickname": f.Nickname,
+			"user_id":  f.Uin,
+			"source":   f.Source,
+		})
+	}
+	return OK(fs)
+}
+
+// CQDeleteUnidirectionalFriend 删除单向好友
+//
+// @route(delete_unidirectional_friend)
+// @rename(uin->user_id)
+func (bot *CQBot) CQDeleteUnidirectionalFriend(uin int64) global.MSG {
+	list, err := bot.Client.GetUnidirectionalFriendList()
+	if err != nil {
+		log.Errorf("获取单向好友列表时出现错误: %v", err)
+		return Failed(100, "API_ERROR", err.Error())
+	}
+	for _, f := range list {
+		if f.Uin == uin {
+			if err = bot.Client.DeleteUnidirectionalFriend(uin); err != nil {
+				log.Errorf("删除单向好友时出现错误: %v", err)
+				return Failed(100, "API_ERROR", err.Error())
+			}
+			return OK(nil)
+		}
+	}
+	return Failed(100, "FRIEND_NOT_FOUND", "好友不存在")
+}
+
 // CQDeleteFriend 删除好友
-//
-//
-func (bot *CQBot) CQDeleteFriend(uin int64) MSG {
+// @route(delete_friend)
+// @rename(uin->"[user_id\x2Cid].0")
+func (bot *CQBot) CQDeleteFriend(uin int64) global.MSG {
 	if bot.Client.FindFriend(uin) == nil {
 		return Failed(100, "FRIEND_NOT_FOUND", "好友不存在")
 	}
@@ -87,13 +300,14 @@ func (bot *CQBot) CQDeleteFriend(uin int64) MSG {
 // CQGetGroupList 获取群列表
 //
 // https://git.io/Jtz1t
-func (bot *CQBot) CQGetGroupList(noCache bool) MSG {
-	gs := make([]MSG, 0, len(bot.Client.GroupList))
+// @route(get_group_list)
+func (bot *CQBot) CQGetGroupList(noCache bool) global.MSG {
+	gs := make([]global.MSG, 0, len(bot.Client.GroupList))
 	if noCache {
 		_ = bot.Client.ReloadGroupList()
 	}
 	for _, g := range bot.Client.GroupList {
-		gs = append(gs, MSG{
+		gs = append(gs, global.MSG{
 			"group_id":          g.Code,
 			"group_name":        g.Name,
 			"group_memo":        g.Memo,
@@ -109,7 +323,8 @@ func (bot *CQBot) CQGetGroupList(noCache bool) MSG {
 // CQGetGroupInfo 获取群信息
 //
 // https://git.io/Jtz1O
-func (bot *CQBot) CQGetGroupInfo(groupID int64, noCache bool) MSG {
+// @route(get_group_info)
+func (bot *CQBot) CQGetGroupInfo(groupID int64, noCache bool) global.MSG {
 	group := bot.Client.FindGroup(groupID)
 	if group == nil || noCache {
 		group, _ = bot.Client.GetGroupInfo(groupID)
@@ -122,7 +337,7 @@ func (bot *CQBot) CQGetGroupInfo(groupID int64, noCache bool) MSG {
 		}
 		for _, g := range info {
 			if g.Code == groupID {
-				return OK(MSG{
+				return OK(global.MSG{
 					"group_id":          g.Code,
 					"group_name":        g.Name,
 					"group_memo":        g.Memo,
@@ -134,7 +349,7 @@ func (bot *CQBot) CQGetGroupInfo(groupID int64, noCache bool) MSG {
 			}
 		}
 	} else {
-		return OK(MSG{
+		return OK(global.MSG{
 			"group_id":          group.Code,
 			"group_name":        group.Name,
 			"group_memo":        group.Memo,
@@ -150,7 +365,8 @@ func (bot *CQBot) CQGetGroupInfo(groupID int64, noCache bool) MSG {
 // CQGetGroupMemberList 获取群成员列表
 //
 // https://git.io/Jtz13
-func (bot *CQBot) CQGetGroupMemberList(groupID int64, noCache bool) MSG {
+// @route(get_group_member_list)
+func (bot *CQBot) CQGetGroupMemberList(groupID int64, noCache bool) global.MSG {
 	group := bot.Client.FindGroup(groupID)
 	if group == nil {
 		return Failed(100, "GROUP_NOT_FOUND", "群聊不存在")
@@ -163,7 +379,7 @@ func (bot *CQBot) CQGetGroupMemberList(groupID int64, noCache bool) MSG {
 		}
 		group.Members = t
 	}
-	members := make([]MSG, 0, len(group.Members))
+	members := make([]global.MSG, 0, len(group.Members))
 	for _, m := range group.Members {
 		members = append(members, convertGroupMemberInfo(groupID, m))
 	}
@@ -173,7 +389,8 @@ func (bot *CQBot) CQGetGroupMemberList(groupID int64, noCache bool) MSG {
 // CQGetGroupMemberInfo 获取群成员信息
 //
 // https://git.io/Jtz1s
-func (bot *CQBot) CQGetGroupMemberInfo(groupID, userID int64, noCache bool) MSG {
+// @route(get_group_member_info)
+func (bot *CQBot) CQGetGroupMemberInfo(groupID, userID int64, noCache bool) global.MSG {
 	group := bot.Client.FindGroup(groupID)
 	if group == nil {
 		return Failed(100, "GROUP_NOT_FOUND", "群聊不存在")
@@ -198,7 +415,8 @@ func (bot *CQBot) CQGetGroupMemberInfo(groupID, userID int64, noCache bool) MSG 
 // CQGetGroupFileSystemInfo 扩展API-获取群文件系统信息
 //
 // https://docs.go-cqhttp.org/api/#%E8%8E%B7%E5%8F%96%E7%BE%A4%E6%96%87%E4%BB%B6%E7%B3%BB%E7%BB%9F%E4%BF%A1%E6%81%AF
-func (bot *CQBot) CQGetGroupFileSystemInfo(groupID int64) MSG {
+// @route(get_group_file_system_info)
+func (bot *CQBot) CQGetGroupFileSystemInfo(groupID int64) global.MSG {
 	fs, err := bot.Client.GetGroupFileSystem(groupID)
 	if err != nil {
 		log.Errorf("获取群 %v 文件系统信息失败: %v", groupID, err)
@@ -210,7 +428,8 @@ func (bot *CQBot) CQGetGroupFileSystemInfo(groupID int64) MSG {
 // CQGetGroupRootFiles 扩展API-获取群根目录文件列表
 //
 // https://docs.go-cqhttp.org/api/#%E8%8E%B7%E5%8F%96%E7%BE%A4%E6%A0%B9%E7%9B%AE%E5%BD%95%E6%96%87%E4%BB%B6%E5%88%97%E8%A1%A8
-func (bot *CQBot) CQGetGroupRootFiles(groupID int64) MSG {
+// @route(get_group_root_files)
+func (bot *CQBot) CQGetGroupRootFiles(groupID int64) global.MSG {
 	fs, err := bot.Client.GetGroupFileSystem(groupID)
 	if err != nil {
 		log.Errorf("获取群 %v 文件系统信息失败: %v", groupID, err)
@@ -221,7 +440,7 @@ func (bot *CQBot) CQGetGroupRootFiles(groupID int64) MSG {
 		log.Errorf("获取群 %v 根目录文件失败: %v", groupID, err)
 		return Failed(100, "FILE_SYSTEM_API_ERROR", err.Error())
 	}
-	return OK(MSG{
+	return OK(global.MSG{
 		"files":   files,
 		"folders": folders,
 	})
@@ -230,7 +449,8 @@ func (bot *CQBot) CQGetGroupRootFiles(groupID int64) MSG {
 // CQGetGroupFilesByFolderID 扩展API-获取群子目录文件列表
 //
 // https://docs.go-cqhttp.org/api/#%E8%8E%B7%E5%8F%96%E7%BE%A4%E5%AD%90%E7%9B%AE%E5%BD%95%E6%96%87%E4%BB%B6%E5%88%97%E8%A1%A8
-func (bot *CQBot) CQGetGroupFilesByFolderID(groupID int64, folderID string) MSG {
+// @route(get_group_files_by_folder)
+func (bot *CQBot) CQGetGroupFilesByFolderID(groupID int64, folderID string) global.MSG {
 	fs, err := bot.Client.GetGroupFileSystem(groupID)
 	if err != nil {
 		log.Errorf("获取群 %v 文件系统信息失败: %v", groupID, err)
@@ -241,7 +461,7 @@ func (bot *CQBot) CQGetGroupFilesByFolderID(groupID int64, folderID string) MSG 
 		log.Errorf("获取群 %v 根目录 %v 子文件失败: %v", groupID, folderID, err)
 		return Failed(100, "FILE_SYSTEM_API_ERROR", err.Error())
 	}
-	return OK(MSG{
+	return OK(global.MSG{
 		"files":   files,
 		"folders": folders,
 	})
@@ -250,12 +470,13 @@ func (bot *CQBot) CQGetGroupFilesByFolderID(groupID int64, folderID string) MSG 
 // CQGetGroupFileURL 扩展API-获取群文件资源链接
 //
 // https://docs.go-cqhttp.org/api/#%E8%8E%B7%E5%8F%96%E7%BE%A4%E6%96%87%E4%BB%B6%E8%B5%84%E6%BA%90%E9%93%BE%E6%8E%A5
-func (bot *CQBot) CQGetGroupFileURL(groupID int64, fileID string, busID int32) MSG {
+// @route(get_group_file_url)
+func (bot *CQBot) CQGetGroupFileURL(groupID int64, fileID string, busID int32) global.MSG {
 	url := bot.Client.GetGroupFileUrl(groupID, fileID, busID)
 	if url == "" {
 		return Failed(100, "FILE_SYSTEM_API_ERROR")
 	}
-	return OK(MSG{
+	return OK(global.MSG{
 		"url": url,
 	})
 }
@@ -263,7 +484,8 @@ func (bot *CQBot) CQGetGroupFileURL(groupID int64, fileID string, busID int32) M
 // CQUploadGroupFile 扩展API-上传群文件
 //
 // https://docs.go-cqhttp.org/api/#%E4%B8%8A%E4%BC%A0%E7%BE%A4%E6%96%87%E4%BB%B6
-func (bot *CQBot) CQUploadGroupFile(groupID int64, file, name, folder string) MSG {
+// @route(upload_group_file)
+func (bot *CQBot) CQUploadGroupFile(groupID int64, file, name, folder string) global.MSG {
 	if !global.PathExists(file) {
 		log.Errorf("上传群文件 %v 失败: 文件不存在", file)
 		return Failed(100, "FILE_NOT_FOUND", "文件不存在")
@@ -285,8 +507,8 @@ func (bot *CQBot) CQUploadGroupFile(groupID int64, file, name, folder string) MS
 
 // CQGroupFileCreateFolder 拓展API-创建群文件文件夹
 //
-//
-func (bot *CQBot) CQGroupFileCreateFolder(groupID int64, parentID, name string) MSG {
+// @route(create_group_file_folder)
+func (bot *CQBot) CQGroupFileCreateFolder(groupID int64, parentID, name string) global.MSG {
 	fs, err := bot.Client.GetGroupFileSystem(groupID)
 	if err != nil {
 		log.Errorf("获取群 %v 文件系统信息失败: %v", groupID, err)
@@ -301,8 +523,9 @@ func (bot *CQBot) CQGroupFileCreateFolder(groupID int64, parentID, name string) 
 
 // CQGroupFileDeleteFolder 拓展API-删除群文件文件夹
 //
-//
-func (bot *CQBot) CQGroupFileDeleteFolder(groupID int64, id string) MSG {
+// @route(delete_group_folder)
+// @rename(id->folder_id)
+func (bot *CQBot) CQGroupFileDeleteFolder(groupID int64, id string) global.MSG {
 	fs, err := bot.Client.GetGroupFileSystem(groupID)
 	if err != nil {
 		log.Errorf("获取群 %v 文件系统信息失败: %v", groupID, err)
@@ -317,8 +540,9 @@ func (bot *CQBot) CQGroupFileDeleteFolder(groupID int64, id string) MSG {
 
 // CQGroupFileDeleteFile 拓展API-删除群文件
 //
-//
-func (bot *CQBot) CQGroupFileDeleteFile(groupID int64, id string, busID int32) MSG {
+// @route(delete_group_file)
+// @rename(id->file_id)
+func (bot *CQBot) CQGroupFileDeleteFile(groupID int64, id string, busID int32) global.MSG {
 	fs, err := bot.Client.GetGroupFileSystem(groupID)
 	if err != nil {
 		log.Errorf("获取群 %v 文件系统信息失败: %v", groupID, err)
@@ -334,7 +558,8 @@ func (bot *CQBot) CQGroupFileDeleteFile(groupID int64, id string, busID int32) M
 // CQGetWordSlices 隐藏API-获取中文分词
 //
 // https://docs.go-cqhttp.org/api/#%E8%8E%B7%E5%8F%96%E4%B8%AD%E6%96%87%E5%88%86%E8%AF%8D-%E9%9A%90%E8%97%8F-api
-func (bot *CQBot) CQGetWordSlices(content string) MSG {
+// @route(.get_word_slices)
+func (bot *CQBot) CQGetWordSlices(content string) global.MSG {
 	slices, err := bot.Client.GetWordSegmentation(content)
 	if err != nil {
 		return Failed(100, "WORD_SEGMENTATION_API_ERROR", err.Error())
@@ -342,13 +567,33 @@ func (bot *CQBot) CQGetWordSlices(content string) MSG {
 	for i := 0; i < len(slices); i++ {
 		slices[i] = strings.ReplaceAll(slices[i], "\u0000", "")
 	}
-	return OK(MSG{"slices": slices})
+	return OK(global.MSG{"slices": slices})
+}
+
+// CQSendMessage 发送消息
+//
+// @route(send_msg)
+// @rename(m->message)
+func (bot *CQBot) CQSendMessage(groupID, userID int64, m gjson.Result, messageType string, autoEscape bool) global.MSG {
+	switch {
+	case messageType == "group":
+		return bot.CQSendGroupMessage(groupID, m, autoEscape)
+	case messageType == "private":
+		fallthrough
+	case userID != 0:
+		return bot.CQSendPrivateMessage(userID, groupID, m, autoEscape)
+	case groupID != 0:
+		return bot.CQSendGroupMessage(groupID, m, autoEscape)
+	}
+	return global.MSG{}
 }
 
 // CQSendGroupMessage 发送群消息
 //
 // https://git.io/Jtz1c
-func (bot *CQBot) CQSendGroupMessage(groupID int64, m gjson.Result, autoEscape bool) MSG {
+// @route(send_group_msg)
+// @rename(m->message)
+func (bot *CQBot) CQSendGroupMessage(groupID int64, m gjson.Result, autoEscape bool) global.MSG {
 	group := bot.Client.FindGroup(groupID)
 	if group == nil {
 		return Failed(100, "GROUP_NOT_FOUND", "群聊不存在")
@@ -366,40 +611,90 @@ func (bot *CQBot) CQSendGroupMessage(groupID int64, m gjson.Result, autoEscape b
 		}
 	}
 
-	if m.Type == gjson.JSON {
-		elem := bot.ConvertObjectMessage(m, true)
-		fixAt(elem)
-		mid := bot.SendGroupMessage(groupID, &message.SendingMessage{Elements: elem})
-		if mid == -1 {
-			return Failed(100, "SEND_MSG_API_ERROR", "请参考 go-cqhttp 端输出")
-		}
-		log.Infof("发送群 %v(%v) 的消息: %v (%v)", group.Name, groupID, limitedString(ToStringMessage(elem, groupID)), mid)
-		return OK(MSG{"message_id": mid})
-	}
-	str := m.String()
-	if str == "" {
-		log.Warn("群消息发送失败: 信息为空.")
-		return Failed(100, "EMPTY_MSG_ERROR", "消息为空")
-	}
 	var elem []message.IMessageElement
-	if autoEscape {
-		elem = append(elem, message.NewText(str))
+	if m.Type == gjson.JSON {
+		elem = bot.ConvertObjectMessage(m, MessageSourceGroup)
 	} else {
-		elem = bot.ConvertStringMessage(str, true)
+		str := m.String()
+		if str == "" {
+			log.Warn("群消息发送失败: 信息为空.")
+			return Failed(100, "EMPTY_MSG_ERROR", "消息为空")
+		}
+		if autoEscape {
+			elem = []message.IMessageElement{message.NewText(str)}
+		} else {
+			elem = bot.ConvertStringMessage(str, MessageSourceGroup)
+		}
 	}
 	fixAt(elem)
 	mid := bot.SendGroupMessage(groupID, &message.SendingMessage{Elements: elem})
 	if mid == -1 {
 		return Failed(100, "SEND_MSG_API_ERROR", "请参考 go-cqhttp 端输出")
 	}
-	log.Infof("发送群 %v(%v) 的消息: %v (%v)", group.Name, groupID, limitedString(str), mid)
-	return OK(MSG{"message_id": mid})
+	log.Infof("发送群 %v(%v) 的消息: %v (%v)", group.Name, groupID, limitedString(m.String()), mid)
+	return OK(global.MSG{"message_id": mid})
+}
+
+// CQSendGuildChannelMessage 发送频道消息
+//
+// @route(send_guild_channel_msg)
+// @rename(m->message)
+func (bot *CQBot) CQSendGuildChannelMessage(guildID, channelID uint64, m gjson.Result, autoEscape bool) global.MSG {
+	guild := bot.Client.GuildService.FindGuild(guildID)
+	if guild == nil {
+		return Failed(100, "GUILD_NOT_FOUND", "频道不存在")
+	}
+	channel := guild.FindChannel(channelID)
+	if channel == nil {
+		return Failed(100, "CHANNEL_NOT_FOUND", "子频道不存在")
+	}
+	if channel.ChannelType != client.ChannelTypeText {
+		log.Warnf("无法发送频道信息: 频道类型错误, 不接受文本信息")
+		return Failed(100, "CHANNEL_NOT_SUPPORTED_TEXT_MSG", "子频道类型错误, 无法发送文本信息")
+	}
+	fixAt := func(elem []message.IMessageElement) {
+		for _, e := range elem {
+			if at, ok := e.(*message.AtElement); ok && at.Target != 0 && at.Display == "" {
+				mem := guild.FindMember(uint64(at.Target))
+				if mem != nil {
+					at.Display = "@" + mem.Nickname
+				} else {
+					at.Display = "@" + strconv.FormatInt(at.Target, 10)
+				}
+			}
+		}
+	}
+
+	var elem []message.IMessageElement
+	if m.Type == gjson.JSON {
+		elem = bot.ConvertObjectMessage(m, MessageSourceGuildChannel)
+	} else {
+		str := m.String()
+		if str == "" {
+			log.Warn("频道发送失败: 信息为空.")
+			return Failed(100, "EMPTY_MSG_ERROR", "消息为空")
+		}
+		if autoEscape {
+			elem = []message.IMessageElement{message.NewText(str)}
+		} else {
+			elem = bot.ConvertStringMessage(str, MessageSourceGuildChannel)
+		}
+	}
+	fixAt(elem)
+	mid := bot.SendGuildChannelMessage(guildID, channelID, &message.SendingMessage{Elements: elem})
+	if mid == "" {
+		return Failed(100, "SEND_MSG_API_ERROR", "请参考 go-cqhttp 端输出")
+	}
+	log.Infof("发送频道 %v(%v) 子频道 %v(%v) 的消息: %v (%v)", guild.GuildName, guild.GuildId, channel.ChannelName, channel.ChannelId, limitedString(m.String()), mid)
+	return OK(global.MSG{"message_id": mid})
 }
 
 // CQSendGroupForwardMessage 扩展API-发送合并转发(群)
 //
 // https://docs.go-cqhttp.org/api/#%E5%8F%91%E9%80%81%E5%90%88%E5%B9%B6%E8%BD%AC%E5%8F%91-%E7%BE%A4
-func (bot *CQBot) CQSendGroupForwardMessage(groupID int64, m gjson.Result) MSG {
+// @route(send_group_forward_msg)
+// @rename(m->messages)
+func (bot *CQBot) CQSendGroupForwardMessage(groupID int64, m gjson.Result) global.MSG {
 	if m.Type != gjson.JSON {
 		return Failed(100)
 	}
@@ -414,7 +709,7 @@ func (bot *CQBot) CQSendGroupForwardMessage(groupID int64, m gjson.Result) MSG {
 		return true
 	})
 
-	var resolveElement = func(elems []message.IMessageElement) []message.IMessageElement {
+	resolveElement := func(elems []message.IMessageElement) []message.IMessageElement {
 		for i, elem := range elems {
 			switch elem.(type) {
 			case *LocalImageElement, *LocalVideoElement:
@@ -437,20 +732,19 @@ func (bot *CQBot) CQSendGroupForwardMessage(groupID int64, m gjson.Result) MSG {
 		ts.Add(time.Second)
 		if e.Get("data.id").Exists() {
 			i := e.Get("data.id").Int()
-			m := bot.GetMessage(int32(i))
+			m, _ := db.GetGroupMessageByGlobalID(int32(i))
 			if m != nil {
-				sender := m["sender"].(message.Sender)
 				return &message.ForwardNode{
-					SenderId:   sender.Uin,
-					SenderName: (&sender).DisplayName(),
+					SenderId:   m.Attribute.SenderUin,
+					SenderName: m.Attribute.SenderName,
 					Time: func() int32 {
-						msgTime := m["time"].(int32)
+						msgTime := m.Attribute.Timestamp
 						if hasCustom && msgTime == 0 {
 							return int32(ts.Unix())
 						}
-						return msgTime
+						return int32(msgTime)
 					}(),
-					Message: resolveElement(bot.ConvertStringMessage(m["message"].(string), true)),
+					Message: resolveElement(bot.ConvertContentMessage(m.Content, MessageSourceGroup)),
 				}
 			}
 			log.Warnf("警告: 引用消息 %v 错误或数据库未开启.", e.Get("data.id").Str)
@@ -489,7 +783,7 @@ func (bot *CQBot) CQSendGroupForwardMessage(groupID int64, m gjson.Result) MSG {
 				}
 			}
 		}
-		content := bot.ConvertObjectMessage(e.Get("data.content"), true)
+		content := bot.ConvertObjectMessage(e.Get("data.content"), MessageSourceGroup)
 		if uin != 0 && name != "" && len(content) > 0 {
 			return &message.ForwardNode{
 				SenderId:   uin,
@@ -521,7 +815,7 @@ func (bot *CQBot) CQSendGroupForwardMessage(groupID int64, m gjson.Result) MSG {
 			log.Warnf("合并转发(群)消息发送失败: 账号可能被风控.")
 			return Failed(100, "SEND_MSG_API_ERROR", "请参考 go-cqhttp 端输出")
 		}
-		return OK(MSG{
+		return OK(global.MSG{
 			"message_id": bot.InsertGroupMessage(ret),
 		})
 	}
@@ -531,38 +825,36 @@ func (bot *CQBot) CQSendGroupForwardMessage(groupID int64, m gjson.Result) MSG {
 // CQSendPrivateMessage 发送私聊消息
 //
 // https://git.io/Jtz1l
-func (bot *CQBot) CQSendPrivateMessage(userID int64, groupID int64, m gjson.Result, autoEscape bool) MSG {
-	if m.Type == gjson.JSON {
-		elem := bot.ConvertObjectMessage(m, false)
-		mid := bot.SendPrivateMessage(userID, groupID, &message.SendingMessage{Elements: elem})
-		if mid == -1 {
-			return Failed(100, "SEND_MSG_API_ERROR", "请参考 go-cqhttp 端输出")
-		}
-		log.Infof("发送好友 %v(%v)  的消息: %v (%v)", userID, userID, limitedString(m.String()), mid)
-		return OK(MSG{"message_id": mid})
-	}
-	str := m.String()
-	if str == "" {
-		return Failed(100, "EMPTY_MSG_ERROR", "消息为空")
-	}
+// @route(send_private_msg)
+// @rename(m->message)
+func (bot *CQBot) CQSendPrivateMessage(userID int64, groupID int64, m gjson.Result, autoEscape bool) global.MSG {
 	var elem []message.IMessageElement
-	if autoEscape {
-		elem = append(elem, message.NewText(str))
+	if m.Type == gjson.JSON {
+		elem = bot.ConvertObjectMessage(m, MessageSourcePrivate)
 	} else {
-		elem = bot.ConvertStringMessage(str, false)
+		str := m.String()
+		if str == "" {
+			return Failed(100, "EMPTY_MSG_ERROR", "消息为空")
+		}
+		if autoEscape {
+			elem = []message.IMessageElement{message.NewText(str)}
+		} else {
+			elem = bot.ConvertStringMessage(str, MessageSourcePrivate)
+		}
 	}
 	mid := bot.SendPrivateMessage(userID, groupID, &message.SendingMessage{Elements: elem})
 	if mid == -1 {
 		return Failed(100, "SEND_MSG_API_ERROR", "请参考 go-cqhttp 端输出")
 	}
-	log.Infof("发送好友 %v(%v)  的消息: %v (%v)", userID, userID, limitedString(str), mid)
-	return OK(MSG{"message_id": mid})
+	log.Infof("发送好友 %v(%v)  的消息: %v (%v)", userID, userID, limitedString(m.String()), mid)
+	return OK(global.MSG{"message_id": mid})
 }
 
 // CQSetGroupCard 设置群名片(群备注)
 //
 // https://git.io/Jtz1B
-func (bot *CQBot) CQSetGroupCard(groupID, userID int64, card string) MSG {
+// @route(set_group_card)
+func (bot *CQBot) CQSetGroupCard(groupID, userID int64, card string) global.MSG {
 	if g := bot.Client.FindGroup(groupID); g != nil {
 		if m := g.FindMember(userID); m != nil {
 			m.EditCard(card)
@@ -575,7 +867,8 @@ func (bot *CQBot) CQSetGroupCard(groupID, userID int64, card string) MSG {
 // CQSetGroupSpecialTitle 设置群组专属头衔
 //
 // https://git.io/Jtz10
-func (bot *CQBot) CQSetGroupSpecialTitle(groupID, userID int64, title string) MSG {
+// @route(set_group_special_title)
+func (bot *CQBot) CQSetGroupSpecialTitle(groupID, userID int64, title string) global.MSG {
 	if g := bot.Client.FindGroup(groupID); g != nil {
 		if m := g.FindMember(userID); m != nil {
 			m.EditSpecialTitle(title)
@@ -588,7 +881,9 @@ func (bot *CQBot) CQSetGroupSpecialTitle(groupID, userID int64, title string) MS
 // CQSetGroupName 设置群名
 //
 // https://git.io/Jtz12
-func (bot *CQBot) CQSetGroupName(groupID int64, name string) MSG {
+// @route(set_group_name)
+// @rename(name->group_name)
+func (bot *CQBot) CQSetGroupName(groupID int64, name string) global.MSG {
 	if g := bot.Client.FindGroup(groupID); g != nil {
 		g.UpdateName(name)
 		return OK(nil)
@@ -599,7 +894,9 @@ func (bot *CQBot) CQSetGroupName(groupID int64, name string) MSG {
 // CQSetGroupMemo 扩展API-发送群公告
 //
 // https://docs.go-cqhttp.org/api/#%E5%8F%91%E9%80%81%E7%BE%A4%E5%85%AC%E5%91%8A
-func (bot *CQBot) CQSetGroupMemo(groupID int64, msg string, img string) MSG {
+// @route(_send_group_notice)
+// @rename(msg->content, img->image)
+func (bot *CQBot) CQSetGroupMemo(groupID int64, msg, img string) global.MSG {
 	if g := bot.Client.FindGroup(groupID); g != nil {
 		if g.SelfPermission() == client.Member {
 			return Failed(100, "PERMISSION_DENIED", "权限不足")
@@ -627,7 +924,9 @@ func (bot *CQBot) CQSetGroupMemo(groupID int64, msg string, img string) MSG {
 // CQSetGroupKick 群组踢人
 //
 // https://git.io/Jtz1V
-func (bot *CQBot) CQSetGroupKick(groupID, userID int64, msg string, block bool) MSG {
+// @route(set_group_kick)
+// @rename(msg->message, block->reject_add_request)
+func (bot *CQBot) CQSetGroupKick(groupID int64, userID int64, msg string, block bool) global.MSG {
 	if g := bot.Client.FindGroup(groupID); g != nil {
 		if m := g.FindMember(userID); m != nil {
 			err := m.Kick(msg, block)
@@ -643,7 +942,9 @@ func (bot *CQBot) CQSetGroupKick(groupID, userID int64, msg string, block bool) 
 // CQSetGroupBan 群组单人禁言
 //
 // https://git.io/Jtz1w
-func (bot *CQBot) CQSetGroupBan(groupID, userID int64, duration uint32) MSG {
+// @route(set_group_ban)
+// @default(duration=1800)
+func (bot *CQBot) CQSetGroupBan(groupID, userID int64, duration uint32) global.MSG {
 	if g := bot.Client.FindGroup(groupID); g != nil {
 		if m := g.FindMember(userID); m != nil {
 			err := m.Mute(duration)
@@ -662,7 +963,9 @@ func (bot *CQBot) CQSetGroupBan(groupID, userID int64, duration uint32) MSG {
 // CQSetGroupWholeBan 群组全员禁言
 //
 // https://git.io/Jtz1o
-func (bot *CQBot) CQSetGroupWholeBan(groupID int64, enable bool) MSG {
+// @route(set_group_whole_ban)
+// @default(enable=true)
+func (bot *CQBot) CQSetGroupWholeBan(groupID int64, enable bool) global.MSG {
 	if g := bot.Client.FindGroup(groupID); g != nil {
 		g.MuteAll(enable)
 		return OK(nil)
@@ -673,7 +976,8 @@ func (bot *CQBot) CQSetGroupWholeBan(groupID int64, enable bool) MSG {
 // CQSetGroupLeave 退出群组
 //
 // https://git.io/Jtz1K
-func (bot *CQBot) CQSetGroupLeave(groupID int64) MSG {
+// @route(set_group_leave)
+func (bot *CQBot) CQSetGroupLeave(groupID int64) global.MSG {
 	if g := bot.Client.FindGroup(groupID); g != nil {
 		g.Quit()
 		return OK(nil)
@@ -684,7 +988,8 @@ func (bot *CQBot) CQSetGroupLeave(groupID int64) MSG {
 // CQGetAtAllRemain 扩展API-获取群 @全体成员 剩余次数
 //
 // https://docs.go-cqhttp.org/api/#%E8%8E%B7%E5%8F%96%E7%BE%A4-%E5%85%A8%E4%BD%93%E6%88%90%E5%91%98-%E5%89%A9%E4%BD%99%E6%AC%A1%E6%95%B0
-func (bot *CQBot) CQGetAtAllRemain(groupID int64) MSG {
+// @route(get_group_at_all_remain)
+func (bot *CQBot) CQGetAtAllRemain(groupID int64) global.MSG {
 	if g := bot.Client.FindGroup(groupID); g != nil {
 		i, err := bot.Client.GetAtAllRemain(groupID)
 		if err != nil {
@@ -698,7 +1003,9 @@ func (bot *CQBot) CQGetAtAllRemain(groupID int64) MSG {
 // CQProcessFriendRequest 处理加好友请求
 //
 // https://git.io/Jtz11
-func (bot *CQBot) CQProcessFriendRequest(flag string, approve bool) MSG {
+// @route(set_friend_add_request)
+// @default(approve=true)
+func (bot *CQBot) CQProcessFriendRequest(flag string, approve bool) global.MSG {
 	req, ok := bot.friendReqCache.Load(flag)
 	if !ok {
 		return Failed(100, "FLAG_NOT_FOUND", "FLAG不存在")
@@ -714,7 +1021,10 @@ func (bot *CQBot) CQProcessFriendRequest(flag string, approve bool) MSG {
 // CQProcessGroupRequest 处理加群请求／邀请
 //
 // https://git.io/Jtz1D
-func (bot *CQBot) CQProcessGroupRequest(flag, subType, reason string, approve bool) MSG {
+// @route(set_group_add_request)
+// @rename(sub_type->"[sub_type\x2Ctype].0")
+// @default(approve=true)
+func (bot *CQBot) CQProcessGroupRequest(flag, subType, reason string, approve bool) global.MSG {
 	msgs, err := bot.Client.GetGroupSystemMessages()
 	if err != nil {
 		log.Errorf("获取群系统消息失败: %v", err)
@@ -758,30 +1068,30 @@ func (bot *CQBot) CQProcessGroupRequest(flag, subType, reason string, approve bo
 // CQDeleteMessage 撤回消息
 //
 // https:// git.io/Jtz1y
-func (bot *CQBot) CQDeleteMessage(messageID int32) MSG {
-	msg := bot.GetMessage(messageID)
-	if msg == nil {
+// @route(delete_msg)
+func (bot *CQBot) CQDeleteMessage(messageID int32) global.MSG {
+	msg, err := db.GetMessageByGlobalID(messageID)
+	if err != nil {
+		log.Warnf("撤回消息时出现错误: %v", err)
 		return Failed(100, "MESSAGE_NOT_FOUND", "消息不存在")
 	}
-	if _, ok := msg["group"]; ok {
-		if msg["internal-id"] == nil {
-			// TODO 撤回临时对话消息
-			log.Warnf("撤回 %v 失败: 无法撤回临时对话消息", messageID)
-			return Failed(100, "CANNOT_RECALL_TEMP_MSG", "无法撤回临时对话消息")
-		}
-		if err := bot.Client.RecallGroupMessage(msg["group"].(int64), msg["message-id"].(int32), msg["internal-id"].(int32)); err != nil {
+	switch o := msg.(type) {
+	case *db.StoredGroupMessage:
+		if err = bot.Client.RecallGroupMessage(o.GroupCode, o.Attribute.MessageSeq, o.Attribute.InternalID); err != nil {
 			log.Warnf("撤回 %v 失败: %v", messageID, err)
 			return Failed(100, "RECALL_API_ERROR", err.Error())
 		}
-	} else {
-		if msg["sender"].(message.Sender).Uin != bot.Client.Uin {
+	case *db.StoredPrivateMessage:
+		if o.Attribute.SenderUin != bot.Client.Uin {
 			log.Warnf("撤回 %v 失败: 好友会话无法撤回对方消息.", messageID)
 			return Failed(100, "CANNOT_RECALL_FRIEND_MSG", "无法撤回对方消息")
 		}
-		if err := bot.Client.RecallPrivateMessage(msg["target"].(int64), int64(msg["time"].(int32)), msg["message-id"].(int32), msg["internal-id"].(int32)); err != nil {
+		if err = bot.Client.RecallPrivateMessage(o.TargetUin, o.Attribute.Timestamp, o.Attribute.MessageSeq, o.Attribute.InternalID); err != nil {
 			log.Warnf("撤回 %v 失败: %v", messageID, err)
 			return Failed(100, "RECALL_API_ERROR", err.Error())
 		}
+	default:
+		return Failed(100, "UNKNOWN_ERROR")
 	}
 	return OK(nil)
 }
@@ -789,7 +1099,9 @@ func (bot *CQBot) CQDeleteMessage(messageID int32) MSG {
 // CQSetGroupAdmin 群组设置管理员
 //
 // https://git.io/Jtz1S
-func (bot *CQBot) CQSetGroupAdmin(groupID, userID int64, enable bool) MSG {
+// @route(set_group_admin)
+// @default(enable=true)
+func (bot *CQBot) CQSetGroupAdmin(groupID, userID int64, enable bool) global.MSG {
 	group := bot.Client.FindGroup(groupID)
 	if group == nil || group.OwnerUin != bot.Client.Uin {
 		return Failed(100, "PERMISSION_DENIED", "群不存在或权限不足")
@@ -811,12 +1123,13 @@ func (bot *CQBot) CQSetGroupAdmin(groupID, userID int64, enable bool) MSG {
 // CQGetVipInfo 扩展API-获取VIP信息
 //
 // https://docs.go-cqhttp.org/api/#%E8%8E%B7%E5%8F%96vip%E4%BF%A1%E6%81%AF
-func (bot *CQBot) CQGetVipInfo(userID int64) MSG {
+// @route(_get_vip_info)
+func (bot *CQBot) CQGetVipInfo(userID int64) global.MSG {
 	vip, err := bot.Client.GetVipInfo(userID)
 	if err != nil {
 		return Failed(100, "VIP_API_ERROR", err.Error())
 	}
-	msg := MSG{
+	msg := global.MSG{
 		"user_id":          vip.Uin,
 		"nickname":         vip.Name,
 		"level":            vip.Level,
@@ -831,11 +1144,13 @@ func (bot *CQBot) CQGetVipInfo(userID int64) MSG {
 // CQGetGroupHonorInfo 获取群荣誉信息
 //
 // https://git.io/Jtz1H
-func (bot *CQBot) CQGetGroupHonorInfo(groupID int64, t string) MSG {
-	msg := MSG{"group_id": groupID}
-	convertMem := func(memList []client.HonorMemberInfo) (ret []MSG) {
+// @route(get_group_honor_info)
+// @rename(t->type)
+func (bot *CQBot) CQGetGroupHonorInfo(groupID int64, t string) global.MSG {
+	msg := global.MSG{"group_id": groupID}
+	convertMem := func(memList []client.HonorMemberInfo) (ret []global.MSG) {
 		for _, mem := range memList {
-			ret = append(ret, MSG{
+			ret = append(ret, global.MSG{
 				"user_id":     mem.Uin,
 				"nickname":    mem.Name,
 				"avatar":      mem.Avatar,
@@ -847,7 +1162,7 @@ func (bot *CQBot) CQGetGroupHonorInfo(groupID int64, t string) MSG {
 	if t == "talkative" || t == "all" {
 		if honor, err := bot.Client.GetGroupHonorInfo(groupID, client.Talkative); err == nil {
 			if honor.CurrentTalkative.Uin != 0 {
-				msg["current_talkative"] = MSG{
+				msg["current_talkative"] = global.MSG{
 					"user_id":   honor.CurrentTalkative.Uin,
 					"nickname":  honor.CurrentTalkative.Name,
 					"avatar":    honor.CurrentTalkative.Avatar,
@@ -888,12 +1203,13 @@ func (bot *CQBot) CQGetGroupHonorInfo(groupID int64, t string) MSG {
 // CQGetStrangerInfo 获取陌生人信息
 //
 // https://git.io/Jtz17
-func (bot *CQBot) CQGetStrangerInfo(userID int64) MSG {
+// @route(get_stranger_info)
+func (bot *CQBot) CQGetStrangerInfo(userID int64) global.MSG {
 	info, err := bot.Client.GetSummaryInfo(userID)
 	if err != nil {
 		return Failed(100, "SUMMARY_API_ERROR", err.Error())
 	}
-	return OK(MSG{
+	return OK(global.MSG{
 		"user_id":  info.Uin,
 		"nickname": info.Nickname,
 		"qid":      info.Qid,
@@ -915,7 +1231,8 @@ func (bot *CQBot) CQGetStrangerInfo(userID int64) MSG {
 // CQHandleQuickOperation 隐藏API-对事件执行快速操作
 //
 // https://git.io/Jtz15
-func (bot *CQBot) CQHandleQuickOperation(context, operation gjson.Result) MSG {
+// @route(".handle_quick_operation")
+func (bot *CQBot) CQHandleQuickOperation(context, operation gjson.Result) global.MSG {
 	postType := context.Get("post_type").Str
 
 	switch postType {
@@ -926,20 +1243,20 @@ func (bot *CQBot) CQHandleQuickOperation(context, operation gjson.Result) MSG {
 		reply := operation.Get("reply")
 
 		if reply.Exists() {
-			autoEscape := global.EnsureBool(operation.Get("auto_escape"), false)
+			autoEscape := param.EnsureBool(operation.Get("auto_escape"), false)
 			at := operation.Get("at_sender").Bool() && !isAnonymous && msgType == "group"
 			if at && reply.IsArray() {
 				// 在 reply 数组头部插入CQ码
-				replySegments := make([]MSG, 0)
-				segments := make([]MSG, 0)
-				segments = append(segments, MSG{
+				replySegments := make([]global.MSG, 0)
+				segments := make([]global.MSG, 0)
+				segments = append(segments, global.MSG{
 					"type": "at",
-					"data": MSG{
+					"data": global.MSG{
 						"qq": context.Get("sender.user_id").Int(),
 					},
 				})
 
-				err := json.UnmarshalFromString(reply.Raw, &replySegments)
+				err := json.Unmarshal(utils.S2B(reply.Raw), &replySegments)
 				if err != nil {
 					log.WithError(err).Warnf("处理 at_sender 过程中发生错误")
 					return Failed(-1, "处理 at_sender 过程中发生错误", err.Error())
@@ -947,13 +1264,13 @@ func (bot *CQBot) CQHandleQuickOperation(context, operation gjson.Result) MSG {
 
 				segments = append(segments, replySegments...)
 
-				modified, err := json.MarshalToString(segments)
+				modified, err := json.Marshal(segments)
 				if err != nil {
 					log.WithError(err).Warnf("处理 at_sender 过程中发生错误")
 					return Failed(-1, "处理 at_sender 过程中发生错误", err.Error())
 				}
 
-				reply = gjson.Parse(modified)
+				reply = gjson.Parse(utils.B2S(modified))
 			} else if at && reply.Type == gjson.String {
 				reply = gjson.Parse(fmt.Sprintf(
 					"\"[CQ:at,qq=%d]%s\"",
@@ -991,10 +1308,10 @@ func (bot *CQBot) CQHandleQuickOperation(context, operation gjson.Result) MSG {
 		reqType := context.Get("request_type").Str
 		if operation.Get("approve").Exists() {
 			if reqType == "friend" {
-				bot.CQProcessFriendRequest(context.Get("flag").Str, operation.Get("approve").Bool())
+				bot.CQProcessFriendRequest(context.Get("flag").String(), operation.Get("approve").Bool())
 			}
 			if reqType == "group" {
-				bot.CQProcessGroupRequest(context.Get("flag").Str, context.Get("sub_type").Str, operation.Get("reason").Str, operation.Get("approve").Bool())
+				bot.CQProcessGroupRequest(context.Get("flag").String(), context.Get("sub_type").Str, operation.Get("reason").Str, operation.Get("approve").Bool())
 			}
 		}
 	}
@@ -1004,27 +1321,42 @@ func (bot *CQBot) CQHandleQuickOperation(context, operation gjson.Result) MSG {
 // CQGetImage 获取图片(修改自OneBot)
 //
 // https://docs.go-cqhttp.org/api/#%E8%8E%B7%E5%8F%96%E5%9B%BE%E7%89%87%E4%BF%A1%E6%81%AF
-func (bot *CQBot) CQGetImage(file string) MSG {
-	if !global.PathExists(path.Join(global.ImagePath, file)) {
-		return Failed(100)
+// @route(get_image)
+func (bot *CQBot) CQGetImage(file string) global.MSG {
+	var b []byte
+	var err error
+	if cache.EnableCacheDB && strings.HasSuffix(file, ".image") {
+		var f []byte
+		f, err = hex.DecodeString(strings.TrimSuffix(file, ".image"))
+		b = cache.Image.Get(f)
 	}
-	b, err := os.ReadFile(path.Join(global.ImagePath, file))
+
+	if b == nil {
+		if !global.PathExists(path.Join(global.ImagePath, file)) {
+			return Failed(100)
+		}
+		b, err = os.ReadFile(path.Join(global.ImagePath, file))
+	}
+
 	if err == nil {
 		r := binary.NewReader(b)
 		r.ReadBytes(16)
-		msg := MSG{
+		msg := global.MSG{
 			"size":     r.ReadInt32(),
 			"filename": r.ReadString(),
 			"url":      r.ReadString(),
 		}
 		local := path.Join(global.CachePath, file+"."+path.Ext(msg["filename"].(string)))
 		if !global.PathExists(local) {
-			f, _ := os.OpenFile(local, os.O_WRONLY|os.O_CREATE|os.O_TRUNC, 0o0644)
 			if body, err := global.HTTPGetReadCloser(msg["url"].(string)); err == nil {
+				f, _ := os.OpenFile(local, os.O_WRONLY|os.O_CREATE|os.O_TRUNC, 0o0644)
 				_, _ = f.ReadFrom(body)
 				_ = body.Close()
+				_ = f.Close()
+			} else {
+				log.Warnf("下载图片 %v 时出现错误: %v", msg["url"], err)
+				return Failed(100, "DOWNLOAD_IMAGE_ERROR", err.Error())
 			}
-			f.Close()
 		}
 		msg["file"] = local
 		return OK(msg)
@@ -1035,7 +1367,27 @@ func (bot *CQBot) CQGetImage(file string) MSG {
 // CQDownloadFile 扩展API-下载文件到缓存目录
 //
 // https://docs.go-cqhttp.org/api/#%E4%B8%8B%E8%BD%BD%E6%96%87%E4%BB%B6%E5%88%B0%E7%BC%93%E5%AD%98%E7%9B%AE%E5%BD%95
-func (bot *CQBot) CQDownloadFile(url string, headers map[string]string, threadCount int) MSG {
+// @route(download_file)
+func (bot *CQBot) CQDownloadFile(url string, headers gjson.Result, threadCount int) global.MSG {
+	h := map[string]string{}
+	if headers.IsArray() {
+		for _, sub := range headers.Array() {
+			str := strings.SplitN(sub.String(), "=", 2)
+			if len(str) == 2 {
+				h[str[0]] = str[1]
+			}
+		}
+	}
+	if headers.Type == gjson.String {
+		lines := strings.Split(headers.String(), "\r\n")
+		for _, sub := range lines {
+			str := strings.SplitN(sub, "=", 2)
+			if len(str) == 2 {
+				h[str[0]] = str[1]
+			}
+		}
+	}
+
 	hash := md5.Sum([]byte(url))
 	file := path.Join(global.CachePath, hex.EncodeToString(hash[:])+".cache")
 	if global.PathExists(file) {
@@ -1044,12 +1396,12 @@ func (bot *CQBot) CQDownloadFile(url string, headers map[string]string, threadCo
 			return Failed(100, "DELETE_FILE_ERROR", err.Error())
 		}
 	}
-	if err := global.DownloadFileMultiThreading(url, file, 0, threadCount, headers); err != nil {
+	if err := global.DownloadFileMultiThreading(url, file, 0, threadCount, h); err != nil {
 		log.Warnf("下载链接 %v 时出现错误: %v", url, err)
 		return Failed(100, "DOWNLOAD_FILE_ERROR", err.Error())
 	}
 	abs, _ := filepath.Abs(file)
-	return OK(MSG{
+	return OK(global.MSG{
 		"file": abs,
 	})
 }
@@ -1057,24 +1409,26 @@ func (bot *CQBot) CQDownloadFile(url string, headers map[string]string, threadCo
 // CQGetForwardMessage 获取合并转发消息
 //
 // https://git.io/Jtz1F
-func (bot *CQBot) CQGetForwardMessage(resID string) MSG {
+// @route(get_forward_msg)
+// @rename(res_id->"[message_id\x2Cid].0")
+func (bot *CQBot) CQGetForwardMessage(resID string) global.MSG {
 	m := bot.Client.GetForwardMessage(resID)
 	if m == nil {
 		return Failed(100, "MSG_NOT_FOUND", "消息不存在")
 	}
-	r := make([]MSG, 0, len(m.Nodes))
+	r := make([]global.MSG, 0, len(m.Nodes))
 	for _, n := range m.Nodes {
-		bot.checkMedia(n.Message)
-		r = append(r, MSG{
-			"sender": MSG{
+		bot.checkMedia(n.Message, 0)
+		r = append(r, global.MSG{
+			"sender": global.MSG{
 				"user_id":  n.SenderId,
 				"nickname": n.SenderName,
 			},
 			"time":    n.Time,
-			"content": ToFormattedMessage(n.Message, 0, false),
+			"content": ToFormattedMessage(n.Message, MessageSource{SourceType: MessageSourceGroup}, false),
 		})
 	}
-	return OK(MSG{
+	return OK(global.MSG{
 		"messages": r,
 	})
 }
@@ -1082,45 +1436,41 @@ func (bot *CQBot) CQGetForwardMessage(resID string) MSG {
 // CQGetMessage 获取消息
 //
 // https://git.io/Jtz1b
-func (bot *CQBot) CQGetMessage(messageID int32) MSG {
-	msg := bot.GetMessage(messageID)
-	if msg == nil {
+// @route(get_msg)
+func (bot *CQBot) CQGetMessage(messageID int32) global.MSG {
+	msg, err := db.GetMessageByGlobalID(messageID)
+	if err != nil {
+		log.Warnf("获取消息时出现错误: %v", err)
 		return Failed(100, "MSG_NOT_FOUND", "消息不存在")
 	}
-	sender := msg["sender"].(message.Sender)
-	gid, isGroup := msg["group"]
-	raw := msg["message"].(string)
-	return OK(MSG{
-		"message_id":  messageID,
-		"real_id":     msg["message-id"],
-		"message_seq": msg["message-id"],
-		"group":       isGroup,
-		"group_id":    gid,
-		"message_type": func() string {
-			if isGroup {
-				return "group"
-			}
-			return "private"
-		}(),
-		"sender": MSG{
-			"user_id":  sender.Uin,
-			"nickname": sender.Nickname,
+	m := global.MSG{
+		"message_id":    msg.GetGlobalID(),
+		"message_id_v2": msg.GetID(),
+		"message_type":  msg.GetType(),
+		"real_id":       msg.GetAttribute().MessageSeq,
+		"message_seq":   msg.GetAttribute().MessageSeq,
+		"group":         msg.GetType() == "group",
+		"sender": global.MSG{
+			"user_id":  msg.GetAttribute().SenderUin,
+			"nickname": msg.GetAttribute().SenderName,
 		},
-		"time":        msg["time"],
-		"raw_message": raw,
-		"message": ToFormattedMessage(bot.ConvertStringMessage(raw, isGroup), func() int64 {
-			if isGroup {
-				return gid.(int64)
-			}
-			return 0
-		}(), false),
-	})
+		"time": msg.GetAttribute().Timestamp,
+	}
+	switch o := msg.(type) {
+	case *db.StoredGroupMessage:
+		m["group_id"] = o.GroupCode
+		m["message"] = ToFormattedMessage(bot.ConvertContentMessage(o.Content, MessageSourceGroup), MessageSource{SourceType: MessageSourceGroup, PrimaryID: uint64(o.GroupCode)}, false)
+	case *db.StoredPrivateMessage:
+		m["message"] = ToFormattedMessage(bot.ConvertContentMessage(o.Content, MessageSourcePrivate), MessageSource{SourceType: MessageSourcePrivate}, false)
+	}
+	return OK(m)
 }
 
 // CQGetGroupSystemMessages 扩展API-获取群文件系统消息
 //
 // https://docs.go-cqhttp.org/api/#%E8%8E%B7%E5%8F%96%E7%BE%A4%E7%B3%BB%E7%BB%9F%E6%B6%88%E6%81%AF
-func (bot *CQBot) CQGetGroupSystemMessages() MSG {
+// @route(get_group_system_msg)
+func (bot *CQBot) CQGetGroupSystemMessages() global.MSG {
 	msg, err := bot.Client.GetGroupSystemMessages()
 	if err != nil {
 		log.Warnf("获取群系统消息失败: %v", err)
@@ -1132,7 +1482,9 @@ func (bot *CQBot) CQGetGroupSystemMessages() MSG {
 // CQGetGroupMessageHistory 获取群消息历史记录
 //
 // https://docs.go-cqhttp.org/api/#%E8%8E%B7%E5%8F%96%E7%BE%A4%E6%B6%88%E6%81%AF%E5%8E%86%E5%8F%B2%E8%AE%B0%E5%BD%95
-func (bot *CQBot) CQGetGroupMessageHistory(groupID int64, seq int64) MSG {
+// @route(get_group_msg_history)
+// @rename(seq->message_seq)
+func (bot *CQBot) CQGetGroupMessageHistory(groupID int64, seq int64) global.MSG {
 	if g := bot.Client.FindGroup(groupID); g == nil {
 		return Failed(100, "GROUP_NOT_FOUND", "群聊不存在")
 	}
@@ -1148,18 +1500,15 @@ func (bot *CQBot) CQGetGroupMessageHistory(groupID int64, seq int64) MSG {
 		log.Warnf("获取群历史消息失败: %v", err)
 		return Failed(100, "MESSAGES_API_ERROR", err.Error())
 	}
-	ms := make([]MSG, 0, len(msg))
+	ms := make([]global.MSG, 0, len(msg))
 	for _, m := range msg {
-		id := m.Id
-		bot.checkMedia(m.Elements)
-		if bot.db != nil {
-			id = bot.InsertGroupMessage(m)
-		}
+		bot.checkMedia(m.Elements, groupID)
+		id := bot.InsertGroupMessage(m)
 		t := bot.formatGroupMessage(m)
 		t["message_id"] = id
 		ms = append(ms, t)
 	}
-	return OK(MSG{
+	return OK(global.MSG{
 		"messages": ms,
 	})
 }
@@ -1167,22 +1516,23 @@ func (bot *CQBot) CQGetGroupMessageHistory(groupID int64, seq int64) MSG {
 // CQGetOnlineClients 扩展API-获取当前账号在线客户端列表
 //
 // https://docs.go-cqhttp.org/api/#%E8%8E%B7%E5%8F%96%E5%BD%93%E5%89%8D%E8%B4%A6%E5%8F%B7%E5%9C%A8%E7%BA%BF%E5%AE%A2%E6%88%B7%E7%AB%AF%E5%88%97%E8%A1%A8
-func (bot *CQBot) CQGetOnlineClients(noCache bool) MSG {
+// @route(get_online_clients)
+func (bot *CQBot) CQGetOnlineClients(noCache bool) global.MSG {
 	if noCache {
 		if err := bot.Client.RefreshStatus(); err != nil {
 			log.Warnf("刷新客户端状态时出现问题 %v", err)
 			return Failed(100, "REFRESH_STATUS_ERROR", err.Error())
 		}
 	}
-	d := make([]MSG, 0, len(bot.Client.OnlineClients))
+	d := make([]global.MSG, 0, len(bot.Client.OnlineClients))
 	for _, oc := range bot.Client.OnlineClients {
-		d = append(d, MSG{
+		d = append(d, global.MSG{
 			"app_id":      oc.AppId,
 			"device_name": oc.DeviceName,
 			"device_kind": oc.DeviceKind,
 		})
 	}
-	return OK(MSG{
+	return OK(global.MSG{
 		"clients": d,
 	})
 }
@@ -1190,22 +1540,27 @@ func (bot *CQBot) CQGetOnlineClients(noCache bool) MSG {
 // CQCanSendImage 检查是否可以发送图片(此处永远返回true)
 //
 // https://git.io/Jtz1N
-func (bot *CQBot) CQCanSendImage() MSG {
-	return OK(MSG{"yes": true})
+// @route(can_send_image)
+func (bot *CQBot) CQCanSendImage() global.MSG {
+	return OK(global.MSG{"yes": true})
 }
 
 // CQCanSendRecord 检查是否可以发送语音(此处永远返回true)
 //
 // https://git.io/Jtz1x
-func (bot *CQBot) CQCanSendRecord() MSG {
-	return OK(MSG{"yes": true})
+// @route(can_send_record)
+func (bot *CQBot) CQCanSendRecord() global.MSG {
+	return OK(global.MSG{"yes": true})
 }
 
 // CQOcrImage 扩展API-图片OCR
 //
 // https://docs.go-cqhttp.org/api/#%E5%9B%BE%E7%89%87-ocr
-func (bot *CQBot) CQOcrImage(imageID string) MSG {
-	img, err := bot.makeImageOrVideoElem(map[string]string{"file": imageID}, false, true)
+// @route(ocr_image)
+// @alias(.ocr_image)
+// @rename(image_id->image)
+func (bot *CQBot) CQOcrImage(imageID string) global.MSG {
+	img, err := bot.makeImageOrVideoElem(map[string]string{"file": imageID}, false, MessageSourceGroup)
 	if err != nil {
 		log.Warnf("load image error: %v", err)
 		return Failed(100, "LOAD_FILE_ERROR", err.Error())
@@ -1221,7 +1576,8 @@ func (bot *CQBot) CQOcrImage(imageID string) MSG {
 // CQSetGroupPortrait 扩展API-设置群头像
 //
 // https://docs.go-cqhttp.org/api/#%E8%AE%BE%E7%BD%AE%E7%BE%A4%E5%A4%B4%E5%83%8F
-func (bot *CQBot) CQSetGroupPortrait(groupID int64, file, cache string) MSG {
+// @route(set_group_portrait)
+func (bot *CQBot) CQSetGroupPortrait(groupID int64, file, cache string) global.MSG {
 	if g := bot.Client.FindGroup(groupID); g != nil {
 		img, err := global.FindFile(file, cache, global.ImagePath)
 		if err != nil {
@@ -1237,7 +1593,9 @@ func (bot *CQBot) CQSetGroupPortrait(groupID int64, file, cache string) MSG {
 // CQSetGroupAnonymousBan 群组匿名用户禁言
 //
 // https://git.io/Jtz1p
-func (bot *CQBot) CQSetGroupAnonymousBan(groupID int64, flag string, duration int32) MSG {
+// @route(set_group_anonymous_ban)
+// @rename(flag->"[anonymous_flag\x2Canonymous.flag].0")
+func (bot *CQBot) CQSetGroupAnonymousBan(groupID int64, flag string, duration int32) global.MSG {
 	if flag == "" {
 		return Failed(100, "INVALID_FLAG", "无效的flag")
 	}
@@ -1260,8 +1618,9 @@ func (bot *CQBot) CQSetGroupAnonymousBan(groupID int64, flag string, duration in
 // CQGetStatus 获取运行状态
 //
 // https://git.io/JtzMe
-func (bot *CQBot) CQGetStatus() MSG {
-	return OK(MSG{
+// @route(get_status)
+func (bot *CQBot) CQGetStatus() global.MSG {
+	return OK(global.MSG{
 		"app_initialized": true,
 		"app_enabled":     true,
 		"plugins_good":    nil,
@@ -1275,19 +1634,15 @@ func (bot *CQBot) CQGetStatus() MSG {
 // CQSetEssenceMessage 扩展API-设置精华消息
 //
 // https://docs.go-cqhttp.org/api/#%E8%AE%BE%E7%BD%AE%E7%B2%BE%E5%8D%8E%E6%B6%88%E6%81%AF
-func (bot *CQBot) CQSetEssenceMessage(messageID int32) MSG {
-	msg := bot.GetMessage(messageID)
-	if msg == nil {
+// @route(set_essence_msg)
+func (bot *CQBot) CQSetEssenceMessage(messageID int32) global.MSG {
+	msg, err := db.GetGroupMessageByGlobalID(messageID)
+	if err != nil {
 		return Failed(100, "MESSAGE_NOT_FOUND", "消息不存在")
 	}
-	if _, ok := msg["group"]; ok {
-		if err := bot.Client.SetEssenceMessage(msg["group"].(int64), msg["message-id"].(int32), msg["internal-id"].(int32)); err != nil {
-			log.Warnf("设置精华消息 %v 失败: %v", messageID, err)
-			return Failed(100, "SET_ESSENCE_MSG_ERROR", err.Error())
-		}
-	} else {
-		log.Warnf("设置精华消息 %v 失败: 非群聊", messageID)
-		return Failed(100, "SET_ESSENCE_MSG_ERROR", "非群聊")
+	if err := bot.Client.SetEssenceMessage(msg.GroupCode, msg.Attribute.MessageSeq, msg.Attribute.InternalID); err != nil {
+		log.Warnf("设置精华消息 %v 失败: %v", messageID, err)
+		return Failed(100, "SET_ESSENCE_MSG_ERROR", err.Error())
 	}
 	return OK(nil)
 }
@@ -1295,19 +1650,15 @@ func (bot *CQBot) CQSetEssenceMessage(messageID int32) MSG {
 // CQDeleteEssenceMessage 扩展API-移出精华消息
 //
 // https://docs.go-cqhttp.org/api/#%E7%A7%BB%E5%87%BA%E7%B2%BE%E5%8D%8E%E6%B6%88%E6%81%AF
-func (bot *CQBot) CQDeleteEssenceMessage(messageID int32) MSG {
-	msg := bot.GetMessage(messageID)
-	if msg == nil {
+// @route(delete_essence_msg)
+func (bot *CQBot) CQDeleteEssenceMessage(messageID int32) global.MSG {
+	msg, err := db.GetGroupMessageByGlobalID(messageID)
+	if err != nil {
 		return Failed(100, "MESSAGE_NOT_FOUND", "消息不存在")
 	}
-	if _, ok := msg["group"]; ok {
-		if err := bot.Client.DeleteEssenceMessage(msg["group"].(int64), msg["message-id"].(int32), msg["internal-id"].(int32)); err != nil {
-			log.Warnf("移出精华消息 %v 失败: %v", messageID, err)
-			return Failed(100, "DEL_ESSENCE_MSG_ERROR", err.Error())
-		}
-	} else {
-		log.Warnf("移出精华消息 %v 失败: 非群聊", messageID)
-		return Failed(100, "DEL_ESSENCE_MSG_ERROR", "非群聊")
+	if err := bot.Client.DeleteEssenceMessage(msg.GroupCode, msg.Attribute.MessageSeq, msg.Attribute.InternalID); err != nil {
+		log.Warnf("删除精华消息 %v 失败: %v", messageID, err)
+		return Failed(100, "SET_ESSENCE_MSG_ERROR", err.Error())
 	}
 	return OK(nil)
 }
@@ -1315,18 +1666,19 @@ func (bot *CQBot) CQDeleteEssenceMessage(messageID int32) MSG {
 // CQGetEssenceMessageList 扩展API-获取精华消息列表
 //
 // https://docs.go-cqhttp.org/api/#%E8%8E%B7%E5%8F%96%E7%B2%BE%E5%8D%8E%E6%B6%88%E6%81%AF%E5%88%97%E8%A1%A8
-func (bot *CQBot) CQGetEssenceMessageList(groupCode int64) MSG {
-	g := bot.Client.FindGroup(groupCode)
+// @route(get_essence_msg_list)
+func (bot *CQBot) CQGetEssenceMessageList(groupID int64) global.MSG {
+	g := bot.Client.FindGroup(groupID)
 	if g == nil {
 		return Failed(100, "GROUP_NOT_FOUND", "群聊不存在")
 	}
-	msgList, err := bot.Client.GetGroupEssenceMsgList(groupCode)
+	msgList, err := bot.Client.GetGroupEssenceMsgList(groupID)
 	if err != nil {
 		return Failed(100, "GET_ESSENCE_LIST_FOUND", err.Error())
 	}
-	list := make([]MSG, 0, len(msgList))
+	list := make([]global.MSG, 0, len(msgList))
 	for _, m := range msgList {
-		msg := MSG{
+		msg := global.MSG{
 			"sender_nick":   m.SenderNick,
 			"sender_time":   m.SenderTime,
 			"operator_time": m.AddDigestTime,
@@ -1334,7 +1686,7 @@ func (bot *CQBot) CQGetEssenceMessageList(groupCode int64) MSG {
 			"sender_id":     m.SenderUin,
 			"operator_id":   m.AddDigestUin,
 		}
-		msg["message_id"] = toGlobalID(groupCode, int32(m.MessageID))
+		msg["message_id"] = db.ToGlobalID(groupID, int32(m.MessageID))
 		list = append(list, msg)
 	}
 	return OK(list)
@@ -1343,8 +1695,9 @@ func (bot *CQBot) CQGetEssenceMessageList(groupCode int64) MSG {
 // CQCheckURLSafely 扩展API-检查链接安全性
 //
 // https://docs.go-cqhttp.org/api/#%E6%A3%80%E6%9F%A5%E9%93%BE%E6%8E%A5%E5%AE%89%E5%85%A8%E6%80%A7
-func (bot *CQBot) CQCheckURLSafely(url string) MSG {
-	return OK(MSG{
+// @route(check_url_safely)
+func (bot *CQBot) CQCheckURLSafely(url string) global.MSG {
+	return OK(global.MSG{
 		"level": bot.Client.CheckUrlSafely(url),
 	})
 }
@@ -1352,12 +1705,13 @@ func (bot *CQBot) CQCheckURLSafely(url string) MSG {
 // CQGetVersionInfo 获取版本信息
 //
 // https://git.io/JtwUs
-func (bot *CQBot) CQGetVersionInfo() MSG {
+// @route(get_version_info)
+func (bot *CQBot) CQGetVersionInfo() global.MSG {
 	wd, _ := os.Getwd()
-	return OK(MSG{
+	return OK(global.MSG{
 		"app_name":                   "go-cqhttp",
-		"app_version":                Version,
-		"app_full_name":              fmt.Sprintf("go-cqhttp-%s_%s_%s-%s", Version, runtime.GOOS, runtime.GOARCH, runtime.Version()),
+		"app_version":                base.Version,
+		"app_full_name":              fmt.Sprintf("go-cqhttp-%s_%s_%s-%s", base.Version, runtime.GOOS, runtime.GOARCH, runtime.Version()),
 		"protocol_version":           "v11",
 		"coolq_directory":            wd,
 		"coolq_edition":              "pro",
@@ -1367,7 +1721,7 @@ func (bot *CQBot) CQGetVersionInfo() MSG {
 		"plugin_build_configuration": "release",
 		"runtime_version":            runtime.Version(),
 		"runtime_os":                 runtime.GOOS,
-		"version":                    Version,
+		"version":                    base.Version,
 		"protocol": func() int {
 			switch client.SystemDeviceInfo.Protocol {
 			case client.IPad:
@@ -1390,19 +1744,20 @@ func (bot *CQBot) CQGetVersionInfo() MSG {
 // CQGetModelShow 获取在线机型
 //
 // https://club.vip.qq.com/onlinestatus/set
-func (bot *CQBot) CQGetModelShow(modelName string) MSG {
-	variants, err := bot.Client.GetModelShow(modelName)
+// @route(_get_model_show)
+func (bot *CQBot) CQGetModelShow(model string) global.MSG {
+	variants, err := bot.Client.GetModelShow(model)
 	if err != nil {
 		return Failed(100, "GET_MODEL_SHOW_API_ERROR", "无法获取在线机型")
 	}
-	a := make([]MSG, 0, len(variants))
+	a := make([]global.MSG, 0, len(variants))
 	for _, v := range variants {
-		a = append(a, MSG{
+		a = append(a, global.MSG{
 			"model_show": v.ModelShow,
 			"need_pay":   v.NeedPay,
 		})
 	}
-	return OK(MSG{
+	return OK(global.MSG{
 		"variants": a,
 	})
 }
@@ -1410,8 +1765,9 @@ func (bot *CQBot) CQGetModelShow(modelName string) MSG {
 // CQSetModelShow 设置在线机型
 //
 // https://club.vip.qq.com/onlinestatus/set
-func (bot *CQBot) CQSetModelShow(modelName string, modelShow string) MSG {
-	err := bot.Client.SetModelShow(modelName, modelShow)
+// @route(_set_model_show)
+func (bot *CQBot) CQSetModelShow(model, modelShow string) global.MSG {
+	err := bot.Client.SetModelShow(model, modelShow)
 	if err != nil {
 		return Failed(100, "SET_MODEL_SHOW_API_ERROR", "无法设置在线机型")
 	}
@@ -1419,29 +1775,38 @@ func (bot *CQBot) CQSetModelShow(modelName string, modelShow string) MSG {
 }
 
 // CQMarkMessageAsRead 标记消息已读
-func (bot *CQBot) CQMarkMessageAsRead(msgID int32) MSG {
-	m := bot.GetMessage(msgID)
-	if m == nil {
+// @route(mark_msg_as_read)
+// @rename(msg_id->message_id)
+func (bot *CQBot) CQMarkMessageAsRead(msgID int32) global.MSG {
+	m, err := db.GetMessageByGlobalID(msgID)
+	if err != nil {
 		return Failed(100, "MSG_NOT_FOUND", "消息不存在")
 	}
-	if _, ok := m["group"]; ok {
-		bot.Client.MarkGroupMessageReaded(m["group"].(int64), int64(m["message-id"].(int32)))
+	switch o := m.(type) {
+	case *db.StoredGroupMessage:
+		bot.Client.MarkGroupMessageReaded(o.GroupCode, int64(o.Attribute.MessageSeq))
 		return OK(nil)
+	case *db.StoredPrivateMessage:
+		bot.Client.MarkPrivateMessageReaded(o.SessionUin, o.Attribute.Timestamp)
 	}
-	if _, ok := m["from-group"]; ok {
-		return Failed(100, "MSG_TYPE_ERROR", "不支持标记临时会话")
-	}
-	bot.Client.MarkPrivateMessageReaded(m["sender"].(*message.Sender).Uin, m["time"].(int64))
+	return OK(nil)
+}
+
+// CQReloadEventFilter 重载事件过滤器
+//
+// @route(reload_event_filter)
+func (bot *CQBot) CQReloadEventFilter(file string) global.MSG {
+	filter.Add(file)
 	return OK(nil)
 }
 
 // OK 生成成功返回值
-func OK(data interface{}) MSG {
-	return MSG{"data": data, "retcode": 0, "status": "ok"}
+func OK(data interface{}) global.MSG {
+	return global.MSG{"data": data, "retcode": 0, "status": "ok"}
 }
 
 // Failed 生成失败返回值
-func Failed(code int, msg ...string) MSG {
+func Failed(code int, msg ...string) global.MSG {
 	m := ""
 	w := ""
 	if len(msg) > 0 {
@@ -1450,47 +1815,7 @@ func Failed(code int, msg ...string) MSG {
 	if len(msg) > 1 {
 		w = msg[1]
 	}
-	return MSG{"data": nil, "retcode": code, "msg": m, "wording": w, "status": "failed"}
-}
-
-func convertGroupMemberInfo(groupID int64, m *client.GroupMemberInfo) MSG {
-	return MSG{
-		"group_id": groupID,
-		"user_id":  m.Uin,
-		"nickname": m.Nickname,
-		"card":     m.CardName,
-		"sex": func() string {
-			if m.Gender == 1 {
-				return "female"
-			} else if m.Gender == 0 {
-				return "male"
-			}
-			// unknown = 0xff
-			return "unknown"
-		}(),
-		"age":               0,
-		"area":              "",
-		"join_time":         m.JoinTime,
-		"last_sent_time":    m.LastSpeakTime,
-		"shut_up_timestamp": m.ShutUpTimestamp,
-		"level":             strconv.FormatInt(int64(m.Level), 10),
-		"role": func() string {
-			switch m.Permission {
-			case client.Owner:
-				return "owner"
-			case client.Administrator:
-				return "admin"
-			case client.Member:
-				return "member"
-			default:
-				return "member"
-			}
-		}(),
-		"unfriendly":        false,
-		"title":             m.SpecialTitle,
-		"title_expire_time": m.SpecialTitleExpireTime,
-		"card_changeable":   false,
-	}
+	return global.MSG{"data": nil, "retcode": code, "msg": m, "wording": w, "status": "failed"}
 }
 
 func limitedString(str string) string {
