@@ -7,9 +7,9 @@ import (
 	"os/exec"
 	"path"
 
-	"github.com/Mrs4s/go-cqhttp/global/codec"
-
 	"github.com/pkg/errors"
+
+	"github.com/Mrs4s/go-cqhttp/internal/base"
 )
 
 // EncoderSilk 将音频编码为Silk
@@ -23,7 +23,7 @@ func EncoderSilk(data []byte) ([]byte, error) {
 	if silkPath := path.Join("data/cache", tempName+".silk"); PathExists(silkPath) {
 		return os.ReadFile(silkPath)
 	}
-	slk, err := codec.EncodeToSilk(data, tempName, true)
+	slk, err := base.EncodeSilk(data, tempName)
 	if err != nil {
 		return nil, errors.Wrap(err, "encode silk failed")
 	}
