@@ -9,6 +9,7 @@ import (
 	"strings"
 	"sync"
 
+	"github.com/mattn/go-colorable"
 	"github.com/sirupsen/logrus"
 )
 
@@ -80,24 +81,16 @@ func (hook *LocalHook) Fire(entry *logrus.Entry) error {
 }
 
 // SetFormatter 设置日志格式
-func (hook *LocalHook) SetFormatter(formatter logrus.Formatter) {
+func (hook *LocalHook) SetFormatter(consoleFormatter, fileFormatter logrus.Formatter) {
 	hook.lock.Lock()
 	defer hook.lock.Unlock()
 
-	if formatter == nil {
-		// 用默认的
-		formatter = &logrus.TextFormatter{DisableColors: true}
-	} else {
-		switch f := formatter.(type) {
-		case *logrus.TextFormatter:
-			textFormatter := f
-			textFormatter.DisableColors = true
-		default:
-			// todo
-		}
-	}
-	logrus.SetFormatter(formatter)
-	hook.formatter = formatter
+	// 支持处理windows平台的console色彩
+	logrus.SetOutput(colorable.NewColorableStdout())
+	// 用于在console写出
+	logrus.SetFormatter(consoleFormatter)
+	// 用于写入文件
+	hook.formatter = fileFormatter
 }
 
 // SetWriter 设置Writer
@@ -115,11 +108,11 @@ func (hook *LocalHook) SetPath(path string) {
 }
 
 // NewLocalHook 初始化本地日志钩子实现
-func NewLocalHook(args interface{}, formatter logrus.Formatter, levels ...logrus.Level) *LocalHook {
+func NewLocalHook(args interface{}, consoleFormatter, fileFormatter logrus.Formatter, levels ...logrus.Level) *LocalHook {
 	hook := &LocalHook{
 		lock: new(sync.Mutex),
 	}
-	hook.SetFormatter(formatter)
+	hook.SetFormatter(consoleFormatter, fileFormatter)
 	hook.levels = append(hook.levels, levels...)
 
 	switch arg := args.(type) {
@@ -177,12 +170,19 @@ func GetLogLevel(level string) []logrus.Level {
 }
 
 // LogFormat specialize for go-cqhttp
-type LogFormat struct{}
+type LogFormat struct {
+	EnableColor bool
+}
 
 // Format implements logrus.Formatter
 func (f LogFormat) Format(entry *logrus.Entry) ([]byte, error) {
 	buf := NewBuffer()
 	defer PutBuffer(buf)
+
+	if f.EnableColor {
+		buf.WriteString(GetLogLevelColorCode(entry.Level))
+	}
+
 	buf.WriteByte('[')
 	buf.WriteString(entry.Time.Format("2006-01-02 15:04:05"))
 	buf.WriteString("] [")
@@ -190,6 +190,45 @@ func (f LogFormat) Format(entry *logrus.Entry) ([]byte, error) {
 	buf.WriteString("]: ")
 	buf.WriteString(entry.Message)
 	buf.WriteString(" \n")
+
+	if f.EnableColor {
+		buf.WriteString(colorReset)
+	}
+
 	ret := append([]byte(nil), buf.Bytes()...) // copy buffer
 	return ret, nil
+}
+
+const (
+	colorCodePanic = "\x1b[1;31m" // color.Style{color.Bold, color.Red}.String()
+	colorCodeFatal = "\x1b[1;31m" // color.Style{color.Bold, color.Red}.String()
+	colorCodeError = "\x1b[31m"   // color.Style{color.Red}.String()
+	colorCodeWarn  = "\x1b[33m"   // color.Style{color.Yellow}.String()
+	colorCodeInfo  = "\x1b[37m"   // color.Style{color.White}.String()
+	colorCodeDebug = "\x1b[32m"   // color.Style{color.Green}.String()
+	colorCodeTrace = "\x1b[36m"   // color.Style{color.Cyan}.String()
+	colorReset     = "\x1b[0m"
+)
+
+// GetLogLevelColorCode 获取日志等级对应色彩code
+func GetLogLevelColorCode(level logrus.Level) string {
+	switch level {
+	case logrus.PanicLevel:
+		return colorCodePanic
+	case logrus.FatalLevel:
+		return colorCodeFatal
+	case logrus.ErrorLevel:
+		return colorCodeError
+	case logrus.WarnLevel:
+		return colorCodeWarn
+	case logrus.InfoLevel:
+		return colorCodeInfo
+	case logrus.DebugLevel:
+		return colorCodeDebug
+	case logrus.TraceLevel:
+		return colorCodeTrace
+
+	default:
+		return colorCodeInfo
+	}
 }
