@@ -12,22 +12,22 @@ import (
 	"strconv"
 	"strings"
 	"sync"
-	"time"
 
-	"github.com/guonaihong/gout"
 	"github.com/pkg/errors"
 
 	"github.com/tidwall/gjson"
+
+	"github.com/Mrs4s/go-cqhttp/internal/base"
 )
 
 var (
 	client = &http.Client{
 		Transport: &http.Transport{
 			Proxy: func(request *http.Request) (u *url.URL, e error) {
-				if Proxy == "" {
+				if base.Proxy == "" {
 					return http.ProxyFromEnvironment(request)
 				}
-				return url.Parse(Proxy)
+				return url.Parse(base.Proxy)
 			},
 			ForceAttemptHTTP2:   true,
 			MaxConnsPerHost:     0,
@@ -35,9 +35,6 @@ var (
 			MaxIdleConnsPerHost: 999,
 		},
 	}
-
-	// Proxy 存储Config.proxy_rewrite,用于设置代理
-	Proxy string
 
 	// ErrOverSize 响应主体过大时返回此错误
 	ErrOverSize = errors.New("oversize")
@@ -247,22 +244,6 @@ func DownloadFileMultiThreading(url, path string, limit int64, threadCount int, 
 	}
 	wg.Wait()
 	return lastErr
-}
-
-// GetSliderTicket 通过给定的验证链接raw和id,获取验证结果Ticket
-func GetSliderTicket(raw, id string) (string, error) {
-	var rsp string
-	if err := gout.POST("https://api.shkong.com/gocqhttpapi/task").SetJSON(gout.H{
-		"id":  id,
-		"url": raw,
-	}).SetTimeout(time.Second * 35).BindBody(&rsp).Do(); err != nil {
-		return "", err
-	}
-	g := gjson.Parse(rsp)
-	if g.Get("error").Str != "" {
-		return "", errors.New(g.Get("error").Str)
-	}
-	return g.Get("ticket").Str, nil
 }
 
 // QQMusicSongInfo 通过给定id在QQ音乐上查找曲目信息
