@@ -1,30 +1,30 @@
-package update
+package selfupdate
 
 import (
-	"archive/zip"
 	"bytes"
 	"crypto/sha256"
 	"errors"
-	"fmt"
 	"io"
 	"net/http"
+
+	"github.com/klauspost/compress/zip"
 )
 
-// Update go-cqhttp自我更新
-func Update(url string, sum []byte) error {
+// update go-cqhttp自我更新
+func update(url string, sum []byte) error {
 	resp, err := http.Get(url)
 	if err != nil {
 		return err
 	}
 	defer resp.Body.Close()
-	wc := WriteSumCounter{
-		Hash: sha256.New(),
+	wc := writeSumCounter{
+		hash: sha256.New(),
 	}
 	rsp, err := io.ReadAll(io.TeeReader(resp.Body, &wc))
 	if err != nil {
 		return err
 	}
-	if !bytes.Equal(wc.Hash.Sum(nil), sum) {
+	if !bytes.Equal(wc.hash.Sum(nil), sum) {
 		return errors.New("文件已损坏")
 	}
 	reader, _ := zip.NewReader(bytes.NewReader(rsp), resp.ContentLength)
@@ -32,8 +32,7 @@ func Update(url string, sum []byte) error {
 	if err != nil {
 		return err
 	}
-	err, _ = FromStream(file)
-	fmt.Println()
+	err, _ = fromStream(file)
 	if err != nil {
 		return err
 	}
