@@ -9,22 +9,22 @@ import (
 
 	qrcodeTerminal "github.com/Baozisoftware/qrcode-terminal-go"
 	"github.com/Mrs4s/MiraiGo/client"
+	"github.com/gocq/qrcode"
 	"github.com/pkg/errors"
 	log "github.com/sirupsen/logrus"
-	"github.com/tuotoo/qrcode"
 
 	"github.com/Mrs4s/go-cqhttp/global"
 )
 
 var console = bufio.NewReader(os.Stdin)
 
-var readLine = func() (str string) {
+func readLine() (str string) {
 	str, _ = console.ReadString('\n')
 	str = strings.TrimSpace(str)
 	return
 }
 
-var readLineTimeout = func(t time.Duration, de string) (str string) {
+func readLineTimeout(t time.Duration, de string) (str string) {
 	r := make(chan string)
 	go func() {
 		select {
@@ -70,7 +70,7 @@ func qrcodeLogin() error {
 		log.Infof("请使用手机QQ扫描二维码 (qrcode.png) : ")
 	}
 	time.Sleep(time.Second)
-	qrcodeTerminal.New().Get(fi.Content).Print()
+	qrcodeTerminal.New2(qrcodeTerminal.ConsoleColors.BrightBlack, qrcodeTerminal.ConsoleColors.BrightWhite, qrcodeTerminal.QRCodeRecoveryLevels.Low).Get(fi.Content).Print()
 	s, err := cli.QueryQRCodeStatus(rsp.Sig)
 	if err != nil {
 		return err
