@@ -3,60 +3,25 @@ package server
 import (
 	"container/list"
 	"context"
-	"os"
 	"sync"
+	"time"
 
 	"github.com/Mrs4s/go-cqhttp/coolq"
 	"github.com/Mrs4s/go-cqhttp/global"
+	"github.com/Mrs4s/go-cqhttp/modules/api"
 
-	log "github.com/sirupsen/logrus"
-	"github.com/tidwall/gjson"
 	"golang.org/x/time/rate"
 )
 
-var (
-	filters     = make(map[string]global.Filter)
-	filterMutex sync.RWMutex
-)
-
-func rateLimit(frequency float64, bucketSize int) handler {
+func rateLimit(frequency float64, bucketSize int) api.Handler {
 	limiter := rate.NewLimiter(rate.Limit(frequency), bucketSize)
-	return func(_ string, _ resultGetter) coolq.MSG {
+	return func(_ string, _ api.Getter) global.MSG {
 		_ = limiter.Wait(context.Background())
 		return nil
 	}
 }
 
-func addFilter(file string) {
-	if file == "" {
-		return
-	}
-	bs, err := os.ReadFile(file)
-	if err != nil {
-		log.Error("init filter error: ", err)
-		return
-	}
-	defer func() {
-		if err := recover(); err != nil {
-			log.Error("init filter error: ", err)
-		}
-	}()
-	filter := global.Generate("and", gjson.ParseBytes(bs))
-	filterMutex.Lock()
-	filters[file] = filter
-	filterMutex.Unlock()
-}
-
-func findFilter(file string) global.Filter {
-	if file == "" {
-		return nil
-	}
-	filterMutex.RLock()
-	defer filterMutex.RUnlock()
-	return filters[file]
-}
-
-func longPolling(bot *coolq.CQBot, maxSize int) handler {
+func longPolling(bot *coolq.CQBot, maxSize int) api.Handler {
 	var mutex sync.Mutex
 	cond := sync.NewCond(&mutex)
 	queue := list.New()
@@ -69,23 +34,43 @@ func longPolling(bot *coolq.CQBot, maxSize int) handler {
 		}
 		cond.Signal()
 	})
-	return func(action string, p resultGetter) coolq.MSG {
+	return func(action string, p api.Getter) global.MSG {
 		if action != "get_updates" {
 			return nil
 		}
-		mutex.Lock()
-		defer mutex.Unlock()
-		if queue.Len() == 0 {
-			cond.Wait()
+		var (
+			once    sync.Once
+			ch      = make(chan []interface{}, 1)
+			timeout = time.Duration(p.Get("timeout").Int()) * time.Second
+		)
+		defer close(ch)
+		go func() {
+			mutex.Lock()
+			defer mutex.Unlock()
+			if queue.Len() == 0 {
+				cond.Wait()
+			}
+			once.Do(func() {
+				limit := int(p.Get("limit").Int())
+				if limit <= 0 || queue.Len() < limit {
+					limit = queue.Len()
+				}
+				ret := make([]interface{}, limit)
+				for i := 0; i < limit; i++ {
+					ret[i] = queue.Remove(queue.Front())
+				}
+				ch <- ret
+			})
+		}()
+		if timeout != 0 {
+			select {
+			case <-time.After(timeout):
+				once.Do(func() {})
+				return coolq.OK([]interface{}{})
+			case ret := <-ch:
+				return coolq.OK(ret)
+			}
 		}
-		limit := int(p.Get("limit").Int())
-		if limit <= 0 || queue.Len() < limit {
-			limit = queue.Len()
-		}
-		ret := make([]interface{}, limit)
-		for i := 0; i < limit; i++ {
-			ret[i] = queue.Remove(queue.Front())
-		}
-		return coolq.OK(ret)
+		return coolq.OK(<-ch)
 	}
 }
